Rename source code files for API servers.
