+++ media-proxy/src/api_server_tcp.cc
@@ -6,7 +6,7 @@
 
 // #include <signal.h>
 
-#include "tcp_controller.h"
+#include "api_server_tcp.h"
 #include "mp_ctrl_proto.h"
 #include <mcm_dp.h>
